adding devices to boot sequence
